Implemented frame skips Setup a system that allows skipping a configurable amounts of screen updates. Can be disabled by setting it to zero or commenting the define. Read below for implications Minor stylistic uniformization Introduced a semi-hardcoded multi-sampling (read below)
After a few tests in an emulator (with SPI nonetheless) I noticed that a reduction of even 3 out of 4 screen delays with user input related changes, causes quite a sense of unresponsiveness. I thus only delayed the redraws of the compass needles. Needs further testing, along with the MULTISAMPLE_SAMPLING.